# src/main/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Crawler API
  version: "1.0"
servers:
- url: https://localhost.com:8080/
paths:
  /crawler/addArticle:
    post:
      operationId: addArticle
      description: Save an Article
      requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
      responses:
          200:
            description: Sucessfully saved an Article
            content:
              'application/json':
                schema:
                  $ref: '#/components/schemas/Article'
          400:
            description: Bad Request
            content:
              'application/json':
                schema:
                  type: object
                  properties:
                    message:
                      type: string
          500:
            description: Internal Server Error
            content:
              'application/json':
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                    
  /crawler/getArticles:
    get:
      operationId: getArticles
      description: Returns an Array of Articles
      responses:
        200:
          description: Sucessfully returned an Array of Articles
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        400:
          description: Bad Request
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
        500:
          description: Internal Server Error
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    
  /crawler/getArticlesByCategory:
    get:
      operationId: getArticlesByCategory
      description: Returns an Array of Articles filtered by Category
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        200:
          description: Sucessfully returned an filtered Array of Articles
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        400:
          description: Bad Request
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
        500:
          description: Internal Server Error
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    
  /crawler/getArticlesByTerm:
    get:
      operationId: getArticlesByTerm
      description: Returns an Array of Articles filtered by Term
      parameters:
        - in: query
          name: term
          schema:
            type: string
      responses:
        200:
          description: Sucessfully returned an filtered Array of Articles
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        400:
          description: Bad Request
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
        500:
          description: Internal Server Error
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string

components:
  schemas:
    Article:
      required:
      - guid
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        guid:
          type: integer
        pubDate:
          type: string
        description:
          type: string
        link:
          type: string